static class 問題の回避実装。
